add custom logic to make the in-SDK gate yaml easier (#13)
This change affects how we author the horton-e2e.yaml which lives in the SDK repo. With this change, we can set Horton.Repo to $(Build.Repository.Uri), which saves us from hardcoding the repo URI into the YAML. This is especially important for C# where we will occasionally need to run Horton builds from repos outside the Azure/ namespace.